+++ packages/moonstone/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
-## [unreleased]
+## [3.0.0-rc.4] - 2019-08-22
 
 ### Changed
 
@@ -11,7 +11,9 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/ContextualPopupDecorator` arrow rendering issue in Chromium
+- `moonstone/EditableIntegerPicker` to properly rerender when the edited value is invalid
 - `moonstone/FormCheckboxItem` to marquee its contents
+- `moonstone/VideoPlayer` to have correct jump forward/backward icon
 - Language-specific fonts so they always use the correct typeface for their locale
 
 ## [3.0.0-rc.3] - 2019-08-15